fix(e2e): split desktop test to own file, fix webview frame locator - W-22888086
The desktop-only test (save query results to CSV/JSON) was included in
the web config, causing electron.launch failures on headless Linux. Move
it to a .desktop.spec.ts file and add testIgnore to the web config.

Also fix the webview frame locator: VS Code desktop-electron renders
webview iframes outside .editor-group-container, so the scoped locator
never found elements. Use page.frameLocator which resolves lazily across
all matching frames at action time.
