+++ packages/salesforcedx-vscode-soql/test/playwright/playwright.config.web.ts
@@ -5,6 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { defineConfig } from '@playwright/test';
 import { createWebConfig } from '@salesforce/playwright-vscode-ext';
 
-export default createWebConfig({ testDir: './specs' });
+export default defineConfig({ ...createWebConfig({ testDir: './specs' }), testIgnore: ['**/*.desktop.spec.ts'] });
